Various fixes
An indent error was causing the fastresume metadata not to be included in multi-file torrents. Also, the program now checks and sanitize filenames in the initial torrent data, not just in the added fastresume metadata. This will fix paths with numerical folder names which were broken by the previous perl script.