Keep file attr if present
Apparently some torrent software stores file attributes in a key in info/files. Because this extra data is in the 'info' key it becomes part of the torrent's infohash, if I remove it the infohash will change. So if the key is present in the original file, it will have to be added to the fastresume file as well to avoid potential tracker issues.